Add tun pseudo device.
The idea is to sandwich a userland program between to tun devices
to manipulate network traffic in some way.
It makes most sense to have this running on the left sender to
cover how the OpenBSD machine handles reception and forwarding
of such packets.
The tun network stack on the OpenBSD machine is currently not tested.